+++ sdk/azidentity/CHANGELOG.md
@@ -14,6 +14,9 @@
 ### Other Changes
 * `AzureCLICredential` imposes its default timeout only when the `Context`
   passed to `GetToken()` has no deadline
+* Added `NewCredentialUnavailableError()`. This function constructs an error indicating
+  a credential can't authenticate and an encompassing `ChainedTokenCredential` should
+  try its next credential, if any.
 
 ## 1.3.0-beta.1 (2022-12-13)